requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in requirements.in
#
alabaster==0.7.16
    # via sphinx
babel==2.14.0
    # via sphinx
build==1.0.3
    # via -r requirements-dev.in
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
coverage[toml]==7.4.0
    # via
    #   coverage
    #   pytest-cov
docutils==0.20.1
    # via sphinx
exceptiongroup==1.2.0
    # via pytest
idna==3.6
    # via requests
imagesize==1.4.1
    # via sphinx
importlib-metadata==7.0.1
    # via
    #   build
    #   sphinx
iniconfig==2.0.0
    # via pytest
jinja2==3.1.3
    # via sphinx
markupsafe==2.1.3
    # via jinja2
packaging==23.2
    # via
    #   build
    #   pytest
    #   sphinx
pbr==6.0.0
    # via sphinxcontrib-apidoc
pluggy==1.3.0
    # via pytest
pockets==0.9.1
    # via sphinxcontrib-napoleon
pygments==2.17.2
    # via sphinx
pyproject-hooks==1.0.0
    # via build
pytest==7.4.4
    # via
    #   -r requirements-dev.in
    #   pytest-cov
pytest-cov==4.1.0
    # via -r requirements-dev.in
python-dateutil==2.8.2
    # via -r requirements.in
requests==2.31.0
    # via sphinx
six==1.16.0
    # via
    #   pockets
    #   python-dateutil
    #   sphinxcontrib-napoleon
snowballstemmer==2.2.0
    # via sphinx
sphinx==7.2.6
    # via
    #   -r requirements-dev.in
    #   sphinxcontrib-apidoc
    #   sphinxcontrib-restbuilder
sphinxcontrib-apidoc==0.4.0
    # via -r requirements-dev.in
sphinxcontrib-applehelp==1.0.8
    # via sphinx
sphinxcontrib-devhelp==1.0.6
    # via sphinx
sphinxcontrib-htmlhelp==2.0.5
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-napoleon==0.7
    # via -r requirements-dev.in
sphinxcontrib-qthelp==1.0.7
    # via sphinx
sphinxcontrib-restbuilder==0.3
    # via -r requirements-dev.in
sphinxcontrib-serializinghtml==1.1.10
    # via sphinx
tomli==2.0.1
    # via
    #   build
    #   coverage
    #   pyproject-hooks
    #   pytest
urllib3==2.1.0
    # via requests
zipp==3.17.0
    # via importlib-metadata
